Modify the way the new TOC hash algorithms are stored in the header
Only the new hash algorithms (i.e. not none, md5 or sha1) are
stored differently and they are now stored by name in a somewhat
backwards-compatible way (older versions will understand the
header but not the new hash values).  This allows any valid
hash function to be used for the TOC hash algorithm (the same as
for the file checksum) instead of only hash names that are in
an internal baked-in table.